eslint for formatting and added flaticon
