+++ public/index.html
@@ -1,19 +1,19 @@
 <!DOCTYPE html>
-<html lang = "en">
-    <head>
-        <title>CherryPicker</title>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <meta name="description" content="Events venues meta search engine" />
-        <meta name="keywords" content="cloud,transfer,file,files,internet,intranet,scan,ftp,sftp">
-        <meta name="robots" content="noindex, nofollow" />
-        <meta http-equiv="cache-control" content="no-cache" />
-        <meta http-equiv="Pragma" content="no-cache" />
-        <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    </head>
-    <body>
-        <main id="root"></main>
-        <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
-      </body>
-
+<html lang="en">
+  <head>
+    <title>CherryPicker</title>
+    <link rel="icon" href="favicon.ico" />
+    <link rel="manifest" href="./manifest.json" />
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta name="description" content="Events venues meta search engine" />
+    <meta name="robots" content="noindex, nofollow" />
+    <meta http-equiv="cache-control" content="no-cache" />
+    <meta http-equiv="Pragma" content="no-cache" />
+    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+  </head>
+  <body>
+    <main id="root"></main>
+    <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
+  </body>
 </html>
