+++ src/config/theme.js
@@ -1,15 +1,14 @@
 export default {
-    colours: {
-        white: '#ffffff',
-        black: '#000000',
-        maroon: '#960F4A',
-        grey: '#C4C4C4'
-    },
-    breakpoint: {
-        widescreen: '1408px',
-        desktop: '1216px',
-        tablet: '1024px',
-        mobile: '768px',
-    }
-
+  colours: {
+    white: '#ffffff',
+    black: '#000000',
+    maroon: '#960F4A',
+    grey: '#C4C4C4',
+  },
+  breakpoint: {
+    widescreen: '1408px',
+    desktop: '1216px',
+    tablet: '1024px',
+    mobile: '768px',
+  },
 };
