--- src/layouts/PortalLayout.js
@@ -1,109 +1,0 @@
-import React from 'react';
-import {Link} from  'react-router-dom';
-import theme from '../config/theme';
-import styled from 'styled-components';
-import {FaHome} from 'react-icons/fa';
-
-const PortalLayoutStyled = styled('div')`
-    display: flex;
-    height: 100%;
-    width: 100%;
-
-    .layout--main {
-        display: flex;
-        flex-direction: column;
-        align-items: stretch;
-        width: 100%;
-
-        .layout--navbar {
-            display: flex;
-            flex-direction: row;
-            width: 100%;
-            align-items: center;
-            padding-bottom: 30px;
-            margin-top: 25px;
-
-            .navbar--item {
-                display: flex;
-                flex-direction: row;
-                align-items: center;
-                padding-left: 800px;
-                //edit here to adjust the nav-item
-                @media only screen and (max-width: ${theme.breakpoint.widescreen}){
-                    padding-left: 700px;
-                }
-                @media only screen and (max-width: ${theme.breakpoint.desktop}){
-                    padding-left: 450px;
-                }
-                @media only screen and (max-width: ${theme.breakpoint.tablet}){
-                    padding-left: 350px;
-                }
-                @media only screen and (max-width: ${theme.breakpoint.mobile}){
-                    padding-left: 120px;
-                }
-
-            }
-
-            .navbar--title {
-                display: flex;
-                flex-direction: row;
-                align-items: center;
-                justify-content: flex-start;
-                height: 30px;
-                padding-left: 80px;
-            }
-        }
-
-        
-
-        .layout--children {
-            flex: 1 1 auto;
-        }
-
-    }
-`
-
-const NavItem = styled(Link)`
-    font-size: 18px;
-    font-weight: bold;
-    color: ${theme.colours.maroon};
-    text-decoration: ${({ selected } = {}) => (selected ? true : 'none')};
-    padding: 0px 0px 0px 40px;
-
-    :hover {
-        color: ${theme.colours.maroon};
-        opacity: 0.8;
-    }
-`
-
-const NavTitle = styled.h1`
-    font-size:  40px;
-    color: ${props => props.color ? theme.colours.maroon : theme.colours.black}
-`
-
-const PortalLayout = ({children, pathname}) => (
-   <PortalLayoutStyled>
-    <div className="layout--main">
-       <div className="layout--navbar">
-       <div className="navbar--title">
-            <NavTitle color> Cherry</NavTitle>
-            <NavTitle >Picker</NavTitle>
-        </div>
-        <div className="navbar--item">
-            <NavItem selected={pathname === '/about'} to= "/about">
-                About us
-            </NavItem>
-            <NavItem selected={pathname === '/bookmark'} to= "/bookmark">
-                Bookmarks
-            </NavItem>
-            <NavItem selected={pathname === '/'} to= "/">
-                <FaHome/>
-            </NavItem>
-        </div>
-       </div>
-       <div className="layout--children">{children}</div>
-   </div>
-   </PortalLayoutStyled>
-);
-
-export default PortalLayout;
