--- src/pages/About.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import PortalLayout from '../layouts/PortalLayout';
-
-const App = ({location: {pathname} = {}} = {}) => (
-    <PortalLayout pathname={pathname}>
-        TODO: About Page
-    </PortalLayout>
-);
-
-export default App;
