+++ src/pages/About.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import PortalLayout from '../layouts/PortalLayout';
+
+const App = ({ location: { pathname } = {} } = {}) => (
+  <PortalLayout pathname={pathname}>TODO: About Page</PortalLayout>
+);
+
+export default App;
