--- src/pages/Bookmark.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import PortalLayout from '../layouts/PortalLayout';
-
-const Bookmark = ({location: {pathname} = {}} = {}) => (
-    <PortalLayout pathname={pathname}>
-        TODO: Bookmark
-    </PortalLayout>
-);
-
-export default Bookmark;
