+++ src/pages/Bookmarks.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import PortalLayout from '../layouts/PortalLayout';
+
+const Bookmarks = ({ location: { pathname } = {} } = {}) => (
+  <PortalLayout pathname={pathname}>TODO: Bookmark</PortalLayout>
+);
+
+export default Bookmarks;
