--- src/pages/Details.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import PortalLayout from '../layouts/PortalLayout';
-
-const Details = ({location: {pathname} = {}} = {}) => (
-    <PortalLayout pathname={pathname}>
-        TODO: Details
-    </PortalLayout>
-);
-
-export default Details;
