+++ src/pages/Details.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import PortalLayout from '../layouts/PortalLayout';
+
+const Details = ({ location: { pathname } = {} } = {}) => (
+  <PortalLayout pathname={pathname}>TODO: Details</PortalLayout>
+);
+
+export default Details;
